Add runtime dependency on doas_cl to 'qdoas' package.
This automatically installs doas_cl when a user installs the qdoas
package, as it used to be before the packages were split up, which is
less confusing for users.  I don't see a big use case for installing
qdoas without doas_cl, as doas_cl only takes up a few MB of extra disk
space.  What matters is the ability to install doas_cl without qdoas,
thereby avoiding the dependency on Qt.